fix(sync): ocasionally marked sites as deleted
