+++ src/hooks/use-fetch-wpcom-sites.tsx
@@ -1,5 +1,6 @@
 import * as Sentry from '@sentry/electron/renderer';
 import { useState, useEffect, useRef, useMemo, useCallback } from 'react';
+import { getIpcApi } from '../lib/get-ipc-api';
 import { useAuth } from './use-auth';
 import { useOffline } from './use-offline';
 
@@ -51,6 +52,101 @@ type SitesEndpointSite = {
 
 type SitesEndpointResponse = {
 	sites: SitesEndpointSite[];
+};
+
+/**
+ * Generate updated site data to be stored in `appdata-v1.json` in three steps:
+ *   1. Update the list of `connectedSites` with fresh data (name, URL, etc)
+ *   2. Find any staging sites that have been added to an already connected site
+ *   3. Find any connected staging sites that have been deleted on WordPress.com
+ *
+ * We treat staging sites differently from production sites because users can't connect staging
+ * sites separately from production sites (they're always connected together). So, while deleted
+ * production sites are still rendered in the UI (with a "deleted" notice), we need to automatically
+ * keep the list of staging sites up-to-date, which is where `stagingSitesToAdd` and
+ * `stagingSitesToDelete` comes in.
+ */
+export const reconcileConnectedSites = (
+	connectedSites: SyncSite[],
+	freshWpComSites: SyncSite[]
+): {
+	updatedConnectedSites: SyncSite[];
+	stagingSitesToAdd: SyncSite[];
+	stagingSitesToDelete: { id: number; localSiteId: string }[];
+} => {
+	const updatedConnectedSites = connectedSites.map( ( connectedSite ): SyncSite => {
+		const site = freshWpComSites.find( ( site ) => site.id === connectedSite.id );
+
+		if ( ! site ) {
+			return {
+				...connectedSite,
+				syncSupport: 'deleted',
+			};
+		}
+
+		return {
+			...connectedSite,
+			name: site.name,
+			url: site.url,
+			syncSupport: site.syncSupport,
+			stagingSiteIds: site.stagingSiteIds,
+		};
+	}, [] );
+
+	const stagingSitesToAdd = connectedSites.flatMap( ( connectedSite ) => {
+		const updatedConnectedSite = updatedConnectedSites.find(
+			( site ) => site.id === connectedSite.id
+		);
+
+		if ( ! updatedConnectedSite?.stagingSiteIds.length ) {
+			return [];
+		}
+
+		const addedStagingSiteIds = updatedConnectedSite.stagingSiteIds.filter(
+			( id ) => ! connectedSite.stagingSiteIds.includes( id )
+		);
+
+		return addedStagingSiteIds.flatMap( ( id ): SyncSite[] => {
+			const freshSite = freshWpComSites.find( ( site ) => site.id === id );
+
+			if ( ! freshSite ) {
+				return [];
+			}
+
+			return [
+				{
+					...freshSite,
+					localSiteId: connectedSite.localSiteId,
+					syncSupport: 'already-connected',
+				},
+			];
+		}, [] );
+	} );
+
+	const stagingSitesToDelete = connectedSites.flatMap( ( connectedSite ) => {
+		const updatedConnectedSite = updatedConnectedSites.find(
+			( site ) => site.id === connectedSite.id
+		);
+
+		if ( ! connectedSite?.stagingSiteIds.length ) {
+			return [];
+		}
+
+		return connectedSite.stagingSiteIds
+			.filter( ( id ) => ! updatedConnectedSite?.stagingSiteIds.includes( id ) )
+			.map( ( id ) => {
+				return {
+					id,
+					localSiteId: connectedSite.localSiteId,
+				};
+			} );
+	} );
+
+	return {
+		updatedConnectedSites,
+		stagingSitesToAdd,
+		stagingSitesToDelete,
+	};
 };
 
 const STUDIO_SYNC_FEATURE_NAME = 'studio-sync';
@@ -120,7 +216,6 @@ export const useFetchWpComSites = ( connectedSiteIds: number[] ) => {
 	const [ rawSyncSites, setRawSyncSites ] = useState< SitesEndpointSite[] >( [] );
 	const { isAuthenticated, client } = useAuth();
 	const isFetchingSites = useRef( false );
-	const isInitialized = useRef( false ); // By default syncSites are always empty array, so this flag helps to determine if we have fetched sites at least once
 	const isOffline = useOffline();
 
 	const joinedConnectedSiteIds = connectedSiteIds.join( ',' );
@@ -142,6 +237,8 @@ export const useFetchWpComSites = ( connectedSiteIds: number[] ) => {
 		isFetchingSites.current = true;
 
 		try {
+			const allConnectedSites = await getIpcApi().getConnectedWpcomSites();
+
 			const response = await client.req.get< SitesEndpointResponse >(
 				{
 					apiNamespace: 'rest/v1.2',
@@ -156,7 +253,34 @@ export const useFetchWpComSites = ( connectedSiteIds: number[] ) => {
 				}
 			);
 
-			isInitialized.current = true;
+			const syncSites = transformSiteResponse(
+				response.sites,
+				allConnectedSites.map( ( { id } ) => id )
+			);
+
+			// whenever array of syncSites changes, we need to update connectedSites to keep them updated with wordpress.com
+			const { updatedConnectedSites, stagingSitesToAdd, stagingSitesToDelete } =
+				reconcileConnectedSites( allConnectedSites, syncSites );
+
+			await getIpcApi().updateConnectedWpcomSites( updatedConnectedSites );
+
+			if ( stagingSitesToDelete.length ) {
+				const data = stagingSitesToDelete.map( ( { id, localSiteId } ) => ( {
+					siteIds: [ id ],
+					localSiteId,
+				} ) );
+
+				await getIpcApi().disconnectWpcomSites( data );
+			}
+
+			if ( stagingSitesToAdd.length ) {
+				const data = stagingSitesToAdd.map( ( site ) => ( {
+					sites: [ site ],
+					localSiteId: site.localSiteId,
+				} ) );
+
+				await getIpcApi().connectWpcomSites( data );
+			}
 
 			setRawSyncSites( response.sites );
 
@@ -178,15 +302,14 @@ export const useFetchWpComSites = ( connectedSiteIds: number[] ) => {
 		return fetchSites();
 	}, [ fetchSites ] );
 
-	const syncSites = useMemo(
+	const syncSitesWithSyncSupportForSelectedSite = useMemo(
 		() => transformSiteResponse( rawSyncSites, memoizedConnectedSiteIds ),
 		[ rawSyncSites, memoizedConnectedSiteIds ]
 	);
 
 	return {
-		syncSites,
+		syncSites: syncSitesWithSyncSupportForSelectedSite,
 		isFetching: isFetchingSites.current,
-		isInitialized: isInitialized.current,
 		refetchSites,
 	};
 };
